ResponseHandler sobre todos los endpoints

+++ src/main/java/com/gruposei/gestion_orquestas/rest/ClothREST.java
@@ -2,6 +2,10 @@ package com.gruposei.gestion_orquestas.rest;
 
 import com.gruposei.gestion_orquestas.model.Cloth;
 import com.gruposei.gestion_orquestas.model.MusicalResource;
+import com.gruposei.gestion_orquestas.model.Role;
+import com.gruposei.gestion_orquestas.model.User;
+import com.gruposei.gestion_orquestas.responses.ApiRequestException;
+import com.gruposei.gestion_orquestas.responses.ResponseHandler;
 import com.gruposei.gestion_orquestas.service.ClothService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -19,35 +23,74 @@ public class ClothREST {
     @Autowired
     private ClothService clothService;
 
+    @Autowired
+    private ResponseHandler responseHandler;
+
     @PostMapping
-    private ResponseEntity<Cloth> save(@RequestBody Cloth p){
+    private ResponseEntity<Object> save(@RequestBody Cloth p){
 
         Cloth temporal = clothService.create(p);
 
         try{
 
-            return ResponseEntity.created((new URI("/api/clothes" + temporal.getId()))).body(temporal);
+            return responseHandler.generateResponse("000",temporal);
         }
         catch(Exception e){
 
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+            throw  new ApiRequestException("002");
         }
     }
 
     @GetMapping
-    private ResponseEntity<List<Cloth>> getAll(){
+    private ResponseEntity<Object> getAll(){
 
-        return ResponseEntity.ok(clothService.getAll());
+        try{
+
+            List<Cloth> cloths = clothService.getAll();
+            return responseHandler.generateResponse("000",cloths);
+        }
+        catch(Exception e){
+
+            throw new ApiRequestException("002");
+        }
     }
 
     @DeleteMapping(params = "id")
-    public ResponseEntity<Void> deleteById(@RequestParam("id") Long id) {
-        clothService.deleteById(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<Object> deleteById(@RequestParam("id") Long id) {
+
+        Optional<Cloth> cloth= clothService.findById(id);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+
+        try {
+            clothService.deleteById(id);
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 
     @RequestMapping(params = "id")
-    public ResponseEntity<Optional<Cloth>> getById(@RequestParam("id") Long id) {
-        return ResponseEntity.ok(clothService.findById(id));
+    public ResponseEntity<Object> getById(@RequestParam("id") Long id) {
+
+        Optional<Cloth> cloth= clothService.findById(id);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+        try {
+
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 }
