+++ src/main/java/com/gruposei/gestion_orquestas/rest/MeetingMinuteREST.java
@@ -2,6 +2,8 @@ package com.gruposei.gestion_orquestas.rest;
 
 import com.gruposei.gestion_orquestas.model.Cloth;
 import com.gruposei.gestion_orquestas.model.MeetingMinute;
+import com.gruposei.gestion_orquestas.responses.ApiRequestException;
+import com.gruposei.gestion_orquestas.responses.ResponseHandler;
 import com.gruposei.gestion_orquestas.service.MeetingMinuteService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -19,35 +21,74 @@ public class MeetingMinuteREST {
     @Autowired
     private MeetingMinuteService meetingMinuteService;
 
+    @Autowired
+    private ResponseHandler responseHandler;
+
     @PostMapping
-    private ResponseEntity<MeetingMinute> save(@RequestBody MeetingMinute p){
+    private ResponseEntity<Object> save(@RequestBody MeetingMinute p){
 
         MeetingMinute temporal = meetingMinuteService.create(p);
 
         try{
 
-            return ResponseEntity.created((new URI("/api/meeting_minutes" + temporal.getId()))).body(temporal);
+            return responseHandler.generateResponse("000",temporal);
         }
         catch(Exception e){
 
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+            throw  new ApiRequestException("002");
         }
     }
 
     @GetMapping
-    private ResponseEntity<List<MeetingMinute>> getAll(){
+    private ResponseEntity<Object> getAll(){
 
-        return ResponseEntity.ok(meetingMinuteService.getAll());
+        try{
+
+            List<MeetingMinute> mm = meetingMinuteService.getAll();
+            return responseHandler.generateResponse("000",mm);
+        }
+        catch(Exception e){
+
+            throw new ApiRequestException("002");
+        }
     }
 
     @DeleteMapping(params = "id")
-    public ResponseEntity<Void> deleteById(@RequestParam("id") Long id) {
-        meetingMinuteService.deleteById(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<Object> deleteById(@RequestParam("id") Long id) {
+
+        Optional<MeetingMinute> cloth= meetingMinuteService.findById(id);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+
+        try {
+            meetingMinuteService.deleteById(id);
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 
     @RequestMapping(params = "id")
-    public ResponseEntity<Optional<MeetingMinute>> getById(@RequestParam("id") Long id) {
-        return ResponseEntity.ok(meetingMinuteService.findById(id));
+    public ResponseEntity<Object> getById(@RequestParam("id") Long id) {
+
+        Optional<MeetingMinute> cloth= meetingMinuteService.findById(id);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+        try {
+
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 }
