+++ src/main/java/com/gruposei/gestion_orquestas/rest/MusicalResourceREST.java
@@ -1,7 +1,10 @@
 package com.gruposei.gestion_orquestas.rest;
 
+import com.gruposei.gestion_orquestas.model.Cloth;
 import com.gruposei.gestion_orquestas.model.MusicalResource;
 import com.gruposei.gestion_orquestas.model.Song;
+import com.gruposei.gestion_orquestas.responses.ApiRequestException;
+import com.gruposei.gestion_orquestas.responses.ResponseHandler;
 import com.gruposei.gestion_orquestas.service.MusicalResourceService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -19,35 +22,74 @@ public class MusicalResourceREST {
     @Autowired
     private MusicalResourceService musicalResourceService;
 
+    @Autowired
+    private ResponseHandler responseHandler;
+
     @PostMapping
-    private ResponseEntity<MusicalResource> save(@RequestBody MusicalResource p){
+    private ResponseEntity<Object> save(@RequestBody MusicalResource p){
 
         MusicalResource temporal = musicalResourceService.create(p);
 
         try{
 
-            return ResponseEntity.created((new URI("/api/musical_resources" + temporal.getId()))).body(temporal);
+            return responseHandler.generateResponse("000",temporal);
         }
         catch(Exception e){
 
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+            throw  new ApiRequestException("002");
         }
     }
 
     @GetMapping
-    private ResponseEntity<List<MusicalResource>> getAll(){
+    private ResponseEntity<Object> getAll(){
 
-        return ResponseEntity.ok(musicalResourceService.getAll());
+        try{
+
+            List<MusicalResource> cloths = musicalResourceService.getAll();
+            return responseHandler.generateResponse("000",cloths);
+        }
+        catch(Exception e){
+
+            throw new ApiRequestException("002");
+        }
     }
 
     @DeleteMapping(params = "id")
-    public ResponseEntity<Void> deleteById(@RequestParam("id") Long id) {
-        musicalResourceService.deleteById(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<Object> deleteById(@RequestParam("id") Long id) {
+
+        Optional<MusicalResource> cloth= musicalResourceService.findById(id);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+
+        try {
+            musicalResourceService.deleteById(id);
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 
     @RequestMapping(params = "id")
-    public ResponseEntity<Optional<MusicalResource>> getById(@RequestParam("id") Long id) {
-        return ResponseEntity.ok(musicalResourceService.findById(id));
+    public ResponseEntity<Object> getById(@RequestParam("id") Long id) {
+
+        Optional<MusicalResource> cloth= musicalResourceService.findById(id);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+        try {
+
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 }
