+++ src/main/java/com/gruposei/gestion_orquestas/rest/MusicalResourceTypeREST.java
@@ -1,6 +1,8 @@
 package com.gruposei.gestion_orquestas.rest;
 
 import com.gruposei.gestion_orquestas.model.MusicalResourceType;
+import com.gruposei.gestion_orquestas.responses.ApiRequestException;
+import com.gruposei.gestion_orquestas.responses.ResponseHandler;
 import com.gruposei.gestion_orquestas.service.MusicalResourceTypeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,40 +20,93 @@ public class MusicalResourceTypeREST {
     @Autowired
     private MusicalResourceTypeService musicalResourceTypeService;
 
+    @Autowired
+    private ResponseHandler responseHandler;
+
     @PostMapping
-    private ResponseEntity<MusicalResourceType> save(@RequestBody MusicalResourceType p){
+    private ResponseEntity<Object> save(@RequestBody MusicalResourceType p){
 
         MusicalResourceType temporal = musicalResourceTypeService.create(p);
 
         try{
 
-            return ResponseEntity.created((new URI("/api/mr_types" + temporal.getId()))).body(temporal);
+            return responseHandler.generateResponse("000",temporal);
         }
         catch(Exception e){
 
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+            throw  new ApiRequestException("002");
         }
     }
 
     @GetMapping
-    private ResponseEntity<List<MusicalResourceType>> getAll(){
+    private ResponseEntity<Object> getAll(){
 
-        return ResponseEntity.ok(musicalResourceTypeService.getAll());
+        try{
+
+            List<MusicalResourceType> cloths = musicalResourceTypeService.getAll();
+            return responseHandler.generateResponse("000",cloths);
+        }
+        catch(Exception e){
+
+            throw new ApiRequestException("002");
+        }
     }
 
     @DeleteMapping(params = "id")
-    public ResponseEntity<Void> deleteById(@RequestParam("id") Long id) {
-        musicalResourceTypeService.deleteById(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<Object> deleteById(@RequestParam("id") Long id) {
+
+        Optional<MusicalResourceType> cloth= musicalResourceTypeService.findById(id);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+
+        try {
+            musicalResourceTypeService.deleteById(id);
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 
     @RequestMapping(params = "id")
-    public ResponseEntity<Optional<MusicalResourceType>> getById(@RequestParam("id") Long id) {
-        return ResponseEntity.ok(musicalResourceTypeService.findById(id));
+    public ResponseEntity<Object> getById(@RequestParam("id") Long id) {
+
+        Optional<MusicalResourceType> cloth= musicalResourceTypeService.findById(id);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+        try {
+
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 
     @RequestMapping(params = "systemName")
-    public ResponseEntity<Optional<MusicalResourceType>> getById(@RequestParam("systemName") String systemName) {
-        return ResponseEntity.ok(musicalResourceTypeService.findBySystemName(systemName));
+    public ResponseEntity<Object> getById(@RequestParam("systemName") String systemName) {
+
+        Optional<MusicalResourceType> cloth = musicalResourceTypeService.findBySystemName(systemName);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+        try {
+
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 }
