+++ src/main/java/com/gruposei/gestion_orquestas/rest/NewContentTypeREST.java
@@ -2,6 +2,8 @@ package com.gruposei.gestion_orquestas.rest;
 
 import com.gruposei.gestion_orquestas.model.MusicalResourceType;
 import com.gruposei.gestion_orquestas.model.NewContentType;
+import com.gruposei.gestion_orquestas.responses.ApiRequestException;
+import com.gruposei.gestion_orquestas.responses.ResponseHandler;
 import com.gruposei.gestion_orquestas.service.NewContentTypeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -19,40 +21,94 @@ public class NewContentTypeREST {
     @Autowired
     private NewContentTypeService newContentTypeService;
 
+    @Autowired
+    private ResponseHandler responseHandler;
+
     @PostMapping
-    private ResponseEntity<NewContentType> save(@RequestBody NewContentType p){
+    private ResponseEntity<Object> save(@RequestBody NewContentType p){
 
         NewContentType temporal = newContentTypeService.create(p);
 
         try{
 
-            return ResponseEntity.created((new URI("/api/nc_types" + temporal.getId()))).body(temporal);
+            return responseHandler.generateResponse("000",temporal);
         }
         catch(Exception e){
 
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+            throw  new ApiRequestException("002");
         }
     }
 
     @GetMapping
-    private ResponseEntity<List<NewContentType>> getAll(){
+    private ResponseEntity<Object> getAll(){
 
-        return ResponseEntity.ok(newContentTypeService.getAll());
+        try{
+
+            List<NewContentType> cloths = newContentTypeService.getAll();
+            return responseHandler.generateResponse("000",cloths);
+        }
+        catch(Exception e){
+
+            throw new ApiRequestException("002");
+        }
     }
 
     @DeleteMapping(params = "id")
-    public ResponseEntity<Void> deleteById(@RequestParam("id") Long id) {
-        newContentTypeService.deleteById(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<Object> deleteById(@RequestParam("id") Long id) {
+
+        Optional<NewContentType> cloth= newContentTypeService.findById(id);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+
+        try {
+            newContentTypeService.deleteById(id);
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 
     @RequestMapping(params = "id")
-    public ResponseEntity<Optional<NewContentType>> getById(@RequestParam("id") Long id) {
-        return ResponseEntity.ok(newContentTypeService.findById(id));
+    public ResponseEntity<Object> getById(@RequestParam("id") Long id) {
+
+        Optional<NewContentType> cloth= newContentTypeService.findById(id);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+
+        try {
+
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 
     @RequestMapping(params = "systemName")
-    public ResponseEntity<Optional<NewContentType>> getById(@RequestParam("systemName") String systemName) {
-        return ResponseEntity.ok(newContentTypeService.findBySystemName(systemName));
+    public ResponseEntity<Object> getById(@RequestParam("systemName") String systemName) {
+
+        Optional<NewContentType> cloth= newContentTypeService.findBySystemName(systemName);
+
+        if(!cloth.isPresent()){
+
+            throw new ApiRequestException("005");
+        }
+        try {
+
+            return responseHandler.generateResponse("000",cloth);
+        }
+        catch (Exception e){
+
+            throw  new ApiRequestException("002");
+        }
     }
 }
